+++ src/main/java/com/aiqin/bms/scmp/api/product/domain/request/allocation/UploadProductRespVo.java
@@ -5,6 +5,7 @@ import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
 
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 
 /**
  * 描述:调拨sku接受实体
@@ -57,11 +58,11 @@ public class UploadProductRespVo {
 
     @ApiModelProperty("税率")
     @NotNull(message = "税率不能为空")
-    private Long tax;
+    private BigDecimal tax;
 
     @ApiModelProperty("含税成本")
     @NotNull(message = "含税成本不能为空")
-    private Long taxPrice;
+    private BigDecimal taxPrice;
 
     @ApiModelProperty("数量")
     @NotNull(message = "数量不能为空")
@@ -69,7 +70,7 @@ public class UploadProductRespVo {
 
     @ApiModelProperty("含税总成本")
     @NotNull(message = "含税总成本不能为空")
-    private Long taxAmount;
+    private BigDecimal taxAmount;
 
     @ApiModelProperty(value = "图片地址",hidden = true)
     private String pictureUrl;
